Split ykshim_client into production and testing modules.
